+++ astore/server/BUILD.bazel
@@ -20,6 +20,7 @@ go_library(
         "//lib/kflags:go_default_library",
         "//lib/kflags/kcobra:go_default_library",
         "//lib/kflags/kconfig:go_default_library",
+        "//lib/khttp/kassets:go_default_library",
         "//lib/khttp/kcookie:go_default_library",
         "//lib/logger:go_default_library",
         "//lib/oauth:go_default_library",
