+++ lib/khttp/kassets/BUILD.bazel
@@ -1,0 +1,13 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["assets.go"],
+    importpath = "github.com/enfabrica/enkit/lib/khttp/kassets",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//lib/khttp:go_default_library",
+        "//lib/logger:go_default_library",
+        "@com_github_dustin_go_humanize//:go_default_library",
+    ],
+)
