+++ lib/oauth/BUILD.bazel
@@ -11,9 +11,9 @@ go_library(
     deps = [
         "//lib/kflags:go_default_library",
         "//lib/khttp:go_default_library",
+        "//lib/khttp/kassets:go_default_library",
         "//lib/khttp/kcookie:go_default_library",
         "//lib/oauth/cookie:go_default_library",
-        "//lib/server:go_default_library",
         "//lib/token:go_default_library",
         "@org_golang_x_oauth2//:go_default_library",
     ],
