+++ lib/oauth/oauth.go
@@ -61,7 +61,7 @@ import (
 	"github.com/enfabrica/enkit/lib/khttp"
 	"github.com/enfabrica/enkit/lib/khttp/kcookie"
 	"github.com/enfabrica/enkit/lib/oauth/cookie"
-	"github.com/enfabrica/enkit/lib/server"
+	"github.com/enfabrica/enkit/lib/khttp/kassets"
 	"github.com/enfabrica/enkit/lib/token"
 )
 
@@ -195,8 +195,8 @@ func (a *Authenticator) LoginURL(target string, state interface{}) (string, []by
 
 // Mapper configures all the URLs to redirect to / unless an authentication cookie is provided by the browser.
 // Further, it configures / to redirect and perform oauth authentication.
-func (auth *Authenticator) Mapper(mapper server.AssetMapper, lm ...LoginModifier) server.AssetMapper {
-	return func(original, name string, handler server.HttpHandler) []string {
+func (auth *Authenticator) Mapper(mapper kassets.AssetMapper, lm ...LoginModifier) kassets.AssetMapper {
+	return func(original, name string, handler khttp.FuncHandler) []string {
 		ext := filepath.Ext(original)
 		switch {
 		case name == "/favicon.ico":
@@ -316,7 +316,7 @@ func (a *Extractor) GetCredentialsFromRequest(r *http.Request) (*CredentialsCook
 //
 // Normally, you should use WithCredentialsOrRedirect(). Use this function only if you
 // expect your handler to be invoked with or without credentials.
-func (a *Extractor) WithCredentials(handler server.HttpHandler) server.HttpHandler {
+func (a *Extractor) WithCredentials(handler khttp.FuncHandler) khttp.FuncHandler {
 	return func(w http.ResponseWriter, r *http.Request) {
 		creds, err := a.GetCredentialsFromRequest(r)
 		if creds != nil && err == nil {
@@ -330,7 +330,7 @@ func (a *Extractor) WithCredentials(handler server.HttpHandler) server.HttpHandl
 //
 // Same as WithCredentials, except that invalid credentials result in a redirect to the specified target.
 // GetCredentials() invoked from the handler is guaranteed to return a non null result.
-func (a *Authenticator) WithCredentialsOrRedirect(handler server.HttpHandler, target string) server.HttpHandler {
+func (a *Authenticator) WithCredentialsOrRedirect(handler khttp.FuncHandler, target string) khttp.FuncHandler {
 	return func(w http.ResponseWriter, r *http.Request) {
 		creds, err := a.GetCredentialsFromRequest(r)
 		if creds == nil || err != nil {
@@ -343,7 +343,7 @@ func (a *Authenticator) WithCredentialsOrRedirect(handler server.HttpHandler, ta
 }
 
 // WithCredentialsOrError invokes the handler if credentials are available, errors out if not.
-func (a *Authenticator) WithCredentialsOrError(handler server.HttpHandler) server.HttpHandler {
+func (a *Authenticator) WithCredentialsOrError(handler khttp.FuncHandler) khttp.FuncHandler {
 	return func(w http.ResponseWriter, r *http.Request) {
 		creds, err := a.GetCredentialsFromRequest(r)
 		if creds == nil || err != nil {
@@ -375,7 +375,7 @@ func (a *Authenticator) WithCredentialsOrError(handler server.HttpHandler) serve
 // with your own code, ensuring it is an absolute URL.
 //
 // Note that login handlers need to be registered with your oauth provider.
-func (a *Authenticator) MakeLoginHandler(handler server.HttpHandler, lm ...LoginModifier) server.HttpHandler {
+func (a *Authenticator) MakeLoginHandler(handler khttp.FuncHandler, lm ...LoginModifier) khttp.FuncHandler {
 	loginHandler := a.LoginHandler(lm...)
 
 	return func(w http.ResponseWriter, r *http.Request) {
@@ -422,7 +422,7 @@ func (a *Authenticator) MakeLoginHandler(handler server.HttpHandler, lm ...Login
 // Note that this call does not allow you to carry any additional state.
 // Use session cookies for that part instead, or get parameters.
 //
-func (a *Authenticator) LoginHandler(lm ...LoginModifier) server.HttpHandler {
+func (a *Authenticator) LoginHandler(lm ...LoginModifier) khttp.FuncHandler {
 	return func(w http.ResponseWriter, r *http.Request) {
 		err := a.PerformLogin(w, r, lm...)
 		if err != nil {
@@ -447,7 +447,7 @@ func (a *Authenticator) LoginHandler(lm ...LoginModifier) server.HttpHandler {
 //
 // Note that auth handlers need to be registered with your oauth provider.
 //
-func (a *Authenticator) MakeAuthHandler(handler server.HttpHandler) server.HttpHandler {
+func (a *Authenticator) MakeAuthHandler(handler khttp.FuncHandler) khttp.FuncHandler {
 	return func(w http.ResponseWriter, r *http.Request) {
 		data, handled, err := a.PerformAuth(w, r)
 		if err == nil && data.Creds != nil {
@@ -476,7 +476,7 @@ func (a *Authenticator) MakeAuthHandler(handler server.HttpHandler) server.HttpH
 // In case of error, an ugly error message is displayed.
 //
 // Use MakeAuthHandler to customize the behavior.
-func (a *Authenticator) AuthHandler() server.HttpHandler {
+func (a *Authenticator) AuthHandler() khttp.FuncHandler {
 	return func(w http.ResponseWriter, r *http.Request) {
 		_, handled, err := a.PerformAuth(w, r)
 		if err != nil {
