+++ lib/server/BUILD.bazel
@@ -2,12 +2,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = ["main.go"],
+    srcs = ["run.go"],
     importpath = "github.com/enfabrica/enkit/lib/server",
     visibility = ["//visibility:public"],
     deps = [
-        "//lib/logger:go_default_library",
-        "@com_github_dustin_go_humanize//:go_default_library",
         "@com_github_improbable_eng_grpc_web//go/grpcweb:go_default_library",
         "@com_github_soheilhy_cmux//:go_default_library",
         "@org_golang_google_grpc//:go_default_library",
